Initial commit of inverse kinematics demo and elliptical orbit comm calculations
Rover inverse kinematics is a demo of a two boned arm. Uses pyplot to show arm state.

Elliptical orbit  is a calculation of the free space loss between a rover and a satellite on an elliptical orbit.
Calculation doesn't account for 3D space such as orbit inclination, but does account for satellite orbital speed variation.